requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
boto3==1.24.62
botocore==1.27.62
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==2.4.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.3
daphne==2.4.1
Deprecated==1.2.13
distlib==0.3.5
Django==4.1.2
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-embed-video==1.4.7
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-widget-tweaks==1.4.12
filelock==3.7.1
Flask==2.2.2
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-connector-python==8.0.30
mysqlclient==2.1.1
nltk==3.7
numpy==1.23.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pandas3==0.0.1
Pillow==9.3.0
platformdirs==2.5.2
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-aiml==0.9.3
python-dateutil==2.8.2
pytz==2021.1
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
skype-chatbot==1.0.4
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tflearn==0.5.0
tqdm==4.64.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==3.10.0.0
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.15.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.1
